reporting cpu/throughput improvements as warning
Removed upper bound for throughput and bottom bound for cpu
measurements as we can say that is improvement and don't want to fail
whole test due some improvements.
Improvement is now evaluated as a pass with warning. E.g. it could be
caused by skipping some part of code but that should be tested by
functional tests.